% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_tickmarks}
\alias{add_tickmarks}
\title{Add tick marks to a time series plot}
\usage{
add_tickmarks(
  obj_lub,
  y_tick_by,
  y_tick_start,
  y_tick_end,
  y_tick_max_number = 50,
  x_tick_extra_years,
  start_decade_ticks
)
}
\arguments{
\item{obj_lub}{A \code{pacea_index} object with a \code{date} column (as a lubridate \code{date}).}

\item{y_tick_by}{Spacing between y-axis tick marks.}

\item{y_tick_start}{Where to start y-axis tick marks.}

\item{y_tick_end}{Where to end y-axis tick marks.}

\item{y_tick_max_number}{Maximum number of y-axis tick marks.}

\item{x_tick_extra_years}{Number of extra years to add tick marks for (does not expand the axis).}

\item{start_decade_ticks}{Year to start decade tick marks (default is 1800).}
}
\value{
Adds tick marks to the current plot. Nothing is returned.
}
\description{
Adds small, unlabeled tick marks to both axes of a time series plot, making it easier to read.
This function is mainly used internally for plots created by the marea and pacea packages.
}
\examples{
\dontrun{
plot.pacea_index(oni)
}
}
\author{
Andrew Edwards
}
