% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{amo}
\alias{amo}
\title{Atlantic MultiDecadal Oscillation (AMO) Index (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{Tibble:
\itemize{
\item{\code{year}:} Year
\item{\code{month}:} Month (1–12)
\item{\code{value}:} SSTA (sea surface temperature anomaly)
}
}
\item{meta}{Includes:
\itemize{
\item{\code{data_type}:} "Atlantic Multidecadal Oscillation Index"
\item{\code{region}:} "Northern Hemisphere (0-60N)"
\item{\code{location_descriptor}:} "North Atlantic"
\item{\code{units}:} ""
\item{\code{source_citation}:} "NOAA , https://www1.ncdc.noaa.gov/pub/data/cmb/ersst/v5/index/ersst.v5.amo.dat"
\item{\code{original_value_col}:} "SSTA"
}
}
}
}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/cmb/ersst/v5/index/ersst.v5.amo.dat}
}
\usage{
amo
}
\description{
Atlantic MultiDecadal Oscillation (AMO) Index (ea_data)
}
\author{
Emily O'Grady
}
\keyword{datasets}
