% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ea_data.R
\name{as_ea_data}
\alias{as_ea_data}
\title{Coerce a pacea object to ea_data (robust multiple-value handling)}
\usage{
as_ea_data(pacea_obj, value_col = NULL, ...)
}
\description{
Coerce a pacea object to ea_data (robust multiple-value handling)
}
