% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ea_seals.R
\name{as_ea_harbour_seals}
\alias{as_ea_harbour_seals}
\title{Coerce a pacea_harbour_seals object to marea's ea_data}
\usage{
as_ea_harbour_seals(pacea_obj)
}
\arguments{
\item{pacea_obj}{A data.frame/tibble from pacea with class 'pacea_harbour_seals'}
}
\value{
An 'ea_data' object.
}
\description{
Adapts a \code{pacea_harbour_seals} object (from pacea) to the \code{ea_data} class.
Handles standard columns (date, region, mean, low, high) and pulls all metadata from attributes.
}
