% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_chemical}
\alias{azmp_chemical}
\title{AZMP chemical data (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character section name CS LL HL BBL
\item{\code{chlorophyll_0_100_value}:} Chlorophyll concentrations integrated between 0 and 100 metres
\item{\code{integrated_nitrate_0_50}:} Nitrate concentrations integrated between 0 and 50 metres
\item{\code{integrated_nitrate_50_150}:} Nitrate concentrations integrated between 50 and 150 metres
\item{\code{integrated_phosphate_0_50}:} Phosphate concentrations integrated between 0 and 50 metres
\item{\code{integrated_phosphate_50_150}:} Phosphate concentrations integrated between 50 and 150 metres
\item{\code{integrated_silicate_0_50s}:} Silicate concentrations integrated between 0 and 50 metres
\item{\code{integrated_silicate_50_150 }:} Silicate concentrations integrated between 50 and 150 metres
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "chemical"
\item{\code{region}:} "Cabot Strait Line, Lousibourg Line, Halifax Line, Browns Bank Line"
\item{\code{location_descriptor}:} "Scotian Shelf"
\item{\code{units}:} "concentrations"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, https://www.bio.gc.ca/science/monitoring-monitorage/azmp-pmza/overview-eng.php"
\item{\code{original_value_col}:} "value"
}
}
}

Casault, B., Johnson, C.L., Devred, E., Clay, S., and Beazley, L. 2025. Chemical and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2024. Can. Tech. Rep. Fish. Aquat. Sci. 3744: vi + 58 p. https://doi.org/10.60825/p6ad-1k52
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdat-bio-load.R}
}
\usage{
azmp_chemical
}
\description{
Ocean chemistry data from AZMP monitoring program CS LL HL BBL Sections
}
\author{
Jamie c. Tam and Benoit Casault
}
\keyword{datasets}
