% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_phytoplankton}
\alias{azmp_phytoplankton}
\title{AZMP phytoplankton data (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character station name for Halifax 2 and Prince 5 where data was collected
\item{\code{diatoms}:} abundance of diatoms (log10)
\item{\code{dinoflagellates}:} abundance of diatoms (log10)
\item{\code{flagillates}:} abundance of diatoms (log10)
\item{\code{cilliates}:} abundance of diatoms (log10)
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "phytoplankton"
\item{\code{region}:} "Halifax2 and Prince5 station"
\item{\code{location_descriptor}:} "Scotian Shelf and Bay of Funday"
\item{\code{units}:} "log10"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, https://www.bio.gc.ca/science/monitoring-monitorage/azmp-pmza/overview-eng.php"
\item{\code{original_value_col}:} "value"
}
}
}

Casault, B., Johnson, C.L., Devred, E., Clay, S., and Beazley, L. 2025. Chemical and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2024. Can. Tech. Rep. Fish. Aquat. Sci. 3744: vi + 58 p. https://doi.org/10.60825/p6ad-1k52
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdat-bio-load.R}
}
\usage{
azmp_phytoplankton
}
\description{
Phytoplankton data from AZMP monitoring program for Halifax 2 and Prince 5 stations
}
\author{
Jamie c. Tam and Benoit Casault
}
\keyword{datasets}
