% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_remotesensing}
\alias{azmp_remotesensing}
\title{AZMP remotesensing data (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character section name CS ESS CSS WSS GB LS
\item{\code{surface_cholorophyll_log10_value}:} Average surface chlorophyll as inferred from satellite (remote sensing) measurements, with a log 10 transform
\item{\code{bloom_start}:} Start of pythoplankton bloom, from satellite
\item{\code{bloom_duration}:} Length of phytoplankton bloom, from satellite
\item{\code{bloom_amplitude}:} Amplitude of phytoplankton bloom, from satellite
\item{\code{bloom_magnitude}:} Magnitude of phytoplankton bloom, from satellite
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "satellite"
\item{\code{region}:} "Cabot Strait, Eastern Scotian Shelf, Central Scotian Shelf, Western Scotian Shelf, Georges Bank, Lower Shelf"
\item{\code{location_descriptor}:} "Scotian Shelf"
\item{\code{units}:} "phytoplankton bloom statistics"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, https://www.bio.gc.ca/science/monitoring-monitorage/azmp-pmza/overview-eng.php"
\item{\code{original_value_col}:} "value"
}
}
}

Casault, B., Johnson, C.L., Devred, E., Clay, S., and Beazley, L. 2025. Chemical and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2024. Can. Tech. Rep. Fish. Aquat. Sci. 3744: vi + 58 p. https://doi.org/10.60825/p6ad-1k52
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdat-bio-load.R}
}
\usage{
azmp_remotesensing
}
\description{
Remote sensing and bloom statistic data from AZMP reporting
}
\author{
Jamie c. Tam and Benoit Casault
}
\keyword{datasets}
