% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_salinity}
\alias{azmp_salinity}
\title{AZMP Surface Salinity (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character NAFO area name
\item{\code{mean_value}:} Numeric surface salinity (psu)
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "Surface Salinity"
\item{\code{region}:} "Scotian Shelf"
\item{\code{location_descriptor}:} "NAFO 4X-4W"
\item{\code{units}:} "psu"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, through azmpdata package https://github.com/casaultb/azmpdata/"
}
}
}

Galbraith, P.S., Lizotte, M., Blais, M., Bélanger, D., Casault, B., Coyne, J., Layton, C., Azetsu-Scott, K., Beazley, L., Chassé, J., Clay, S., Cyr, F., Devred, E., Fudge, A., Gabriel, C.-E., Greenan, B., Hébert, A.-J., Johnson, C.L., Maillet, G., Penney, J., Rastin, S., Ringuette, M., Shaw, J.-L., Snook, S., Starr, M. 2025. Oceanographic conditions in the Atlantic zone in 2024. Can. Tech. Rep. Hydrogr. Ocean. Sci. 400 : viii + 49 p. https://doi.org/10.60825/e92v-d229
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdata-load.R}
}
\usage{
azmp_salinity
}
\description{
Surface salinity from the Atlantic Zone Monitoring Program (AZMP) for the Scotian Shelf, covering NAFO Divisions 4X, 4V, and 4W. Data are averaged from in situ surveys and are crucial for fisheries and climate analyses.
}
\author{
Emily O'Grady and Benoit Casault
}
\keyword{datasets}
