% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_surface_temperature}
\alias{azmp_surface_temperature}
\title{AZMP Scotian Shelf Surface Temperature (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character area name
\item{\code{mean_value}:} Numeric temperature at sea surface (°C)
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "Surface Temperature"
\item{\code{region}:} "Scotian Shelf"
\item{\code{location_descriptor}:} "NAFO 4X-4W"
\item{\code{units}:} "°C"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, through azmpdata package https://github.com/casaultb/azmpdata/"
}
}
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Galbraith, P.S., Lizotte, M., Blais, M., Bélanger, D., Casault, B., Coyne, J., Layton, C., Azetsu-Scott, K., Beazley, L., Chassé, J., Clay, S., Cyr, F., Devred, E., Fudge, A., Gabriel, C.-E., Greenan, B., Hébert, A.-J., Johnson, C.L., Maillet, G., Penney, J., Rastin, S., Ringuette, M., Shaw, J.-L., Snook, S., Starr, M. 2025. Oceanographic conditions in the Atlantic zone in 2024. Can. Tech. Rep. Hydrogr. Ocean. Sci. 400 : viii + 49 p. https://doi.org/10.60825/e92v-d229
 
 
 @docType data
 @name azmp_surface_temperature
 @author Emily O'Grady and Benoit Casault
 @source Generated from running `data-raw/azmpdata/azmpdata-load.R`
}\if{html}{\out{</div>}}
}
\usage{
azmp_surface_temperature
}
\description{
Surface water temperature (annual mean) from the Atlantic Zone Monitoring Program (AZMP) for the Scotian Shelf, covering NAFO Divisions 4X, 4V, and 4W. Data are averaged from in situ surveys and are crucial for fisheries and climate analyses.
}
\keyword{datasets}
