% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_zooplankton}
\alias{azmp_zooplankton}
\title{AZMP zooplankton data (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character section name CS LL HL BBL
\item{\code{calanus_fin_value}:} abundance of Calanus finmarchicus (log10)
\item{\code{pseudocalanus}:} abundance of pseudocalanus (log10)
\item{\code{copepod}:} abundance of copepods group (log10)
\item{\code{non_copepods}:} abundance of non_copepods group (log10)
\item{\code{arctic_calanus_spp}:} abundance arctic calanus group (log10)
\item{\code{warm_offshore_copepods}:} abundance of warm offshore copepods group (Clausocalanus spp., Mecynocera clausi, and Pleuromamma borealis) (log10)
\item{\code{warm_shelf_copepods}:} abundance of warm shelf copepods (the summer-fall copepods Paracalanus spp. and Centropages typicus) (log10)
\item{\code{meso_zooplankton}:} Dry weight of meso zooplankton
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "zooplankton"
\item{\code{region}:} "Cabot Strait Line, Lousibourg Line, Halifax Line, Browns Bank Line"
\item{\code{location_descriptor}:} "Scotian Shelf"
\item{\code{units}:} "log10"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, https://www.bio.gc.ca/science/monitoring-monitorage/azmp-pmza/overview-eng.php"
\item{\code{original_value_col}:} "value"
}
}
}

Casault, B., Johnson, C.L., Devred, E., Clay, S., and Beazley, L. 2025. Chemical and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2024. Can. Tech. Rep. Fish. Aquat. Sci. 3744: vi + 58 p. https://doi.org/10.60825/p6ad-1k52
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdat-bio-load.R}
}
\usage{
azmp_zooplankton
}
\description{
Zooplankton data from AZMP monitoring program CS LL HL BBL Sections
}
\author{
Jamie c. Tam and Benoit Casault
}
\keyword{datasets}
