% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{cite_marea}
\alias{cite_marea}
\title{Get citation information for marea}
\usage{
cite_marea(bibtex = FALSE)
}
\arguments{
\item{bibtex}{Logical. If TRUE, return BibTeX format}
}
\value{
Citation information
}
\description{
Get citation information for marea
}
\examples{
# Get text citation
cite_marea()

# Get BibTeX
cite_marea(bibtex = TRUE)
}
