% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{Coastline Data for North Atlantic Region}
\format{
An \code{sf} object with 72 features and 4 fields:
\describe{
\item{admin}{Character. Country name (e.g., "Canada", "United States of America")}
\item{name_en}{Character. State/province name in English}
\item{name_fr}{Character. State/province name in French}
\item{geometry}{sfc_MULTIPOLYGON. Spatial geometry of the state/province boundaries}
}
}
\source{
Natural Earth Data via rnaturalearth package, see data-raw/coastline/coastline.R
}
\usage{
data(coastline)
}
\description{
State and province level boundaries from Natural Earth for Canada, United States,
Greenland, and Saint Pierre and Miquelon. Useful for adding coastline context to
spatial plots of marine data in the Northwest Atlantic region.
}
\details{
This dataset includes state/province boundaries for:
\itemize{
\item Canada (CA) - All provinces and territories
\item United States (US) - States bordering Atlantic coast
\item Greenland (GL)
\item Saint Pierre and Miquelon (PM)
}

The data is in WGS 84 coordinate reference system (EPSG:4326) and is suitable
for use with ggplot2 and sf plotting functions.
}
\references{
\url{https://www.naturalearthdata.com/}
}
\author{
marea/DFO
}
\keyword{datasets}
