% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_print.R
\name{ea.print}
\alias{ea.print}
\title{Print an ea_data or ea_spatial Object}
\usage{
ea.print(x, ...)
}
\arguments{
\item{x}{an ea_data or ea_spatial object.}

\item{...}{further arguments passed to print method}
}
\description{
Print an ea_data or ea_spatial Object
}
\author{
Emily O'Grady
}
