% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_data_class.R, R/ea_spatial_class.R
\name{ea.subset}
\alias{ea.subset}
\title{Subset an \code{ea_data} Object by Column and Value(s)}
\usage{
ea.subset(x, column, value)

ea.subset(x, column, value)
}
\arguments{
\item{x}{An \code{ea_spatial} object.}

\item{column}{A character string naming the column or layer in the \code{data} slot to filter by.}

\item{value}{A value or vector of values to match in the specified column/layer.}
}
\value{
A new \code{ea_data} object containing only the matching rows.

A new \code{ea_spatial} object containing only the matching rows or cells.
}
\description{
Filters the \code{data} slot of an \code{ea_data} object based on a column and one or more values.

Filters the \code{data} slot of an \code{ea_spatial} object based on a column or layer
and one or more values.
}
\examples{
\dontrun{
ea.subset(azmp_bottom_temperature, "year", 1953)
ea.subset(azmp_bottom_temperature, "region", c("Scotian Shelf", "Gulf"))
}

\dontrun{
ea.subset(temp_obj, "region", "Maritimes")
ea.subset(temp_obj, "value", c(5.5, 5.6))
}

}
