% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_summary.R
\name{ea.summary}
\alias{ea.summary}
\title{Summary method for an ea_data or ea_spatial object}
\usage{
ea.summary(x, ...)
}
\arguments{
\item{x}{an ea_data or ea_spatial object.}

\item{...}{further arguments passed to summary method}
}
\description{
Summary method for an ea_data or ea_spatial object
}
\author{
Emily O'Grady
}
