% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_data_class.R
\name{ea_data-validity}
\alias{ea_data-validity}
\title{Validity Method for ea_data}
\arguments{
\item{object}{An \code{ea_data} object.}
}
\value{
TRUE if valid, otherwise a character string describing the issue.
}
\description{
Ensures that the \code{data} slot contains both \code{year} and \code{value} columns.
}
\keyword{internal}
