% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_spatial_class.R
\name{ea_spatial-validity}
\alias{ea_spatial-validity}
\title{Spatial Data Validity}
\arguments{
\item{object}{An \code{ea_spatial} object.}
}
\value{
\code{TRUE} if valid, otherwise a character vector of error messages.
}
\description{
The validity method for the \code{ea_spatial} class ensures that:
\enumerate{
\item The \code{data} slot contains a supported spatial object (\code{sf}, \code{stars}, \code{SpatRaster}).
\item The spatial object contains a \code{value} column or layer.
}
}
\keyword{internal}
