% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{eco_indicators}
\alias{eco_indicators}
\title{Ecological Indicators (ea_data)}
\format{
An \code{ea_data} object with two components:
\describe{
\item{data}{A tibble with columns:
\describe{
\item{year}{Year}
\item{region}{NAFO division or SS area}
\item{SpeciesRichness_ALL_value}{Species Richness from samples collected in RV survey}
\item{ShannonDiveristy_ALL_value}{Index considering species richness and evenness of abundances}
\item{MargalefRichness_ALL_value}{An index measure of richness accounting for sample size}
\item{PielouEvenness_ALL_value}{How evenly species are distributed within a community}
\item{HillDiversity_ALL_value}{Effective number of species}
\item{HillDominance_ALL_value}{Community dominance index describing how much diversity is driven by a few dominant species}
\item{Heips_ALL_value}{Relative abundance of different species in a community}
\item{ABUNDANCE_ALL_value}{Total abundance}
\item{BIOMASS_ALL_value}{Total biomass of all species caught in RV survey (t)}
\item{BIOMASS_CLUPEIDS_value}{Total biomass of clupeid species (t)}
\item{BIOMASS_FINFISH_value}{Total biomass of finfish species (t)}
\item{BIOMASS_FLATFISH_value}{Total biomass of flatfish species (t)}
\item{BIOMASS_FORAGE_value}{Total biomass of forage species (t)}
\item{BIOMASS_GADOIDS_value}{Total biomass of gadoid species (t)}
\item{BIOMASS_GROUNDFISH_value}{Total biomass of groundfish species (t)}
\item{BIOMASS_PELAGIC_value}{Total biomass of pelagic species (t)}
\item{BIOMASS_SKATES_value}{Total biomass of skate species (t)}
\item{FishinginBalance_value}{Changes in fishing strategies and their impact on system productivity; positive FiB indicates fishery has expanded or bottom-up effects are occurring}
\item{DiversityTargetSpp_ALL_value}{Diversity of target commercial species}
\item{MeanTL.Landings_value}{Mean Trophic level of all landings}
\item{MTI.Landings_3.25_value}{Mean Trophic level of landings above TL 3.25}
\item{landings_ALL_value}{Total landings (t)}
\item{landings_CLUPEIDS.L_value}{Total landings of clupeid species (t)}
\item{landings_FINFISH.L_value}{Total landings of finfish species (t)}
\item{landings_FLATFISH.L_value}{Total landings of flatfish species(t)}
\item{landings_FORAGE.L_value}{Total landings of forage species (t)}
\item{landings_GADOIDS.L_value}{Total landings of gadoid species (t)}
\item{landings_GROUNDFISH.L_value}{Total landings of groundfish species (t)}
\item{landings_INVERTEBRATES.L_value}{Total landings of invertebrate species (t)}
\item{landings_LARGE_PELAGIC.L_value}{Total landings of pelagic species (t)}
\item{landings_SKATES.L_value}{Total landings of skate species (t)}
\item{FP_ALL_value}{Fishing pressure of all species}
\item{FP_CLUPEIDS_value}{Fishing pressure of clupeid species}
\item{FP_FINFISH_value}{Fishing pressure of finfish species}
\item{FP_FLATFISH_value}{Fishing pressure of flatfish species}
\item{FP_FORAGE_value}{Fishing pressure of forage species}
\item{FP_GADOIDS_value}{Fishing pressure of gadoid species}
\item{FP_GROUNDFISH_value}{Fishing pressure of groundfish species}
\item{FP_SKATES_value}{Fishing pressure of skate species}
\item{CVBiomass_value}{Coefficient of variation of the biomass}
\item{MeanLifespan_value}{Mean lifespan observed for each species}
\item{BIOMASS_TL2_value}{Biomass of trophic level 2}
\item{BIOMASS_TL3_value}{Biomass of trophic level 3}
\item{BIOMASS_TL4_value}{Biomass of trophic level 4}
\item{MMLength_BIOMASS_value}{Mean maximum length in the community weighted by biomass}
\item{MMLength_ABUNDANCE_value}{Mean maximum length in the community weighted by abundance}
\item{IVILandings_value}{Intrinsic vulnerability index of the catch}
\item{BIOMASS_LBENTHIVORE_value}{Biomass of large benthivores (t) calculated from RV survey}
\item{BIOMASS_MBENTHIVORE_value}{Biomass of medium-size benthivores (t) calculated from RV survey}
\item{BIOMASS_PISCIVORE_value}{Biomass of piscivores (t) calculated from RV survey}
\item{BIOMASS_PLANKTIVORE_value}{Biomass of planktivores (t) calculated from RV survey}
\item{BIOMASS_ZOOPISCIVORE_value}{Biomass of zoopiscivores (t) calculated from RV survey}
\item{PELAGIC_GROUNDFISH_value}{Biomass of pelagics to biomass of groundfish}
\item{PREDATORS_ALL_value}{Total biomass of predators}
\item{LargeFishIndicator_value}{Proportion of large fish (> 35cm) to small fish (<= 35cm)}
\item{MeanLengthBIOMASS_value}{Mean length of fish in the community weighted by biomass}
\item{MeanLengthABUNDANCE_value}{Mean length of fish in the community weighted by abundance}
\item{CCondition_FINFISH_value}{Community condition of finfish}
\item{CCondition_LBENTHIVORE_value}{Community condition of large benthivores}
\item{CCondition_MBENTHIVORE_value}{Community condition of medium benthivores}
\item{CCondition_PISCIVORE_value}{Community condition of piscivores}
\item{CCondition_PLANKTIVORE_value}{Community condition of planktivores}
\item{CCondition_ZOOPISCIVORE_value}{Community condition of zoopiscivores}
\item{BIOMASS_INVERTEBRATES_value}{Biomass of invertebrates}
\item{INVERTEBRATES_GROUNDFISH_value}{Proportion of invertebrates to groundfish}
\item{FP_INVERTEBRATES_value}{Fishing pressure of invertebrates}
}
}
\item{meta}{A list with elements:
\describe{
\item{data_type}{"ecological"}
\item{region}{"Maritimes"}
\item{location_descriptor}{"NAFO"}
\item{units}{"tonnes"}
\item{species}{"NA"}
\item{source_citation}{"Bundy et al. 2017"}
}
}
}
}
\source{
Generated from running \code{data-raw/eco-indicators/eco-indicators.R}
}
\usage{
eco_indicators
}
\description{
Ecological indicators calculated from biomass estimates and commercial harvesting data.
}
\author{
Jamie C. Tam
}
\keyword{datasets}
