% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{grey_seals}
\alias{grey_seals}
\alias{grey_seals_2021}
\title{Estimated Abundance of Grey Seals on the Scotian Shelf}
\format{
A tibble of class \code{pacea_index} with columns:
\describe{
\item{year}{Year of the estimate}
\item{low}{Lower 95 percent credible interval}
\item{median}{Median estimate of abundance}
\item{high}{Upper 95 percent credible interval}
}

An object of class \code{pacea_biomass} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 62 rows and 4 columns.
}
\source{
Estimates provided by Nell den Heyer. Processed using \code{data-raw/grey-seals/grey-seals/grey-seals.R}.
}
\usage{
grey_seals

grey_seals_2021
}
\description{
Estimates of the abundance of grey seals in Canadian waters, assessed as a single stock.
Coastwide abundance is based on aerial surveys conducted over multiple years and corrected for survey timing, area coverage,
and the proportion of seals hauled out. Estimates are provided for seven regions as well as the coastwide total.
}
\details{
Abundance estimates use Generalized Additive Models (GAMs). The final year of data for each region is saved in
\code{harbour_seals_data_final_year}. The \code{grey_seals} object represents the most recent assessment results.
The \code{grey_seals_2021} object contains results from the 2021 assessment, allowing for reproducibility.

For further details, see Hammill et al. (2023).
}
\examples{
grey_seals
plot(grey_seals)

}
\references{
Hammill, M.O., Rossi, S.P., Mosnier, A., den Heyer, C.E., Bowen, W.D., & Stenson, G.B. (2023).
Grey Seal Abundance and Harvest Advice in Canadian Waters.
DFO Can. Sci. Advis. Sec. Res. Doc. 2023/053. vi + 40 p.
}
\author{
Andrew Edwards, Nell den Heyer
}
\keyword{datasets}
