% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lubridate_pacea_series}
\alias{lubridate_pacea_series}
\title{Convert an object that represents a time seris (\code{pacea_index} etc.) to have lubridate
year/month column, and smooth over a year if required TODO copying over for now.}
\usage{
lubridate_pacea_series(obj, smooth_over_year = FALSE)
}
\arguments{
\item{obj}{A \code{pacea_index} object (a time series).}

\item{smooth_over_year}{Logical. If TRUE, average monthly values over each year.}
}
\value{
original object with month values smoothed (averaged) over the year if
requested, and a new \code{date} column in lubridate format. Sets a resulting
date to be the first of the month or Jan 1st of that year (as appropriate),
for easy
plotting, but note that these should not be taken as exact dates, which is
why we saved the original objects just as year and/or month, not lubridate
dates.
}
\description{
Called from \code{plot.pacea_index()} and others to give a lubrdiate column for
easier plotting.
}
\examples{
\dontrun{
lubridate_pacea_series(oni)
lubridate_pacea_series(oni, smooth_over_year=TRUE)

}
}
\author{
Andrew Edwards
}
