% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marea_citation_reminder}
\alias{marea_citation_reminder}
\title{Print a reminder to cite marea}
\usage{
marea_citation_reminder()
}
\description{
Prints a message reminding users to cite the marea package in their work.
}
\examples{
marea_citation_reminder()
}
