% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{nao}
\alias{nao}
\title{North Atlantic Oscillation (NAO) Index}
\format{
A tibble of class \code{pacea_index} with columns:
\describe{
\item{year}{Year of value}
\item{anomaly}{Index value (strength of westerly winds)}
}
}
\source{
Generated from running \code{data-raw/coastwide-indices/coastwide-indices.R}.
}
\usage{
nao
}
\description{
The NAO Index measures the strength of westerly winds over the North Atlantic, based on the difference in sea-level atmospheric pressures between the Azores and Iceland. It is calculated using a rotated principal component analysis of monthly standardized 500 mb height anomalies, averaged over December to March.
}
\details{
Data are from the \href{https://www.ncei.noaa.gov/access/monitoring/nao/}{NOAA NAO monitoring site}.
}
\examples{
\dontrun{
nao
plot(nao)
}
}
\author{
Jamie C. Tam; Chantelle Layton
}
\keyword{datasets}
