% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{nlci}
\alias{nlci}
\title{Newfoundland and Labrador Climate Index (ea_data)}
\format{
An \code{ea_data} object with:
\describe{
\item{data}{A tibble:
\describe{
\item{year}{year (numeric)}
\item{nlci_value}{nlci_value (numeric)}
}
}
\item{meta}{A named list with:
\describe{
\item{data_type}{"NL Climate Index"}
\item{region}{ "Newfoundland and Labrador Shelf"}
\item{location_descriptor}{"Newfoundland and Labrador Shelf"}
\item{units}{""}
\item{species}{"subindices: Winter North Atlantic Oscillation, air temperature, sea ice, icebergs, sea surface temperature, vertically-averaged temperature and salinity at Station~27, cold intermediate layer (CIL) core temperature at Station~27, CIL area on 3 hydrographic sections and bottom temperature on the NL shelf"}
\item{source_citation}{Character. "Cyr, F. and Galbraith, P. S. 2021. A climate index for the Newfoundland and Labrador shelf, Earth Syst. Sci. Data, 13, 1807–1828, https://doi.org/10.5194/essd-13-1807-2021"}
\item{original_value_col}{Character. "nlci"}
\item{comment}{Character. "Data and associated sources are from https://doi.org/10.20383/101.0301"}
}
}
}
}
\source{
\url{https://doi.org/10.20383/101.0301}
}
\usage{
nlci
}
\description{
The Newfoundland and Labrador (NL) climate index aims to describe the environmental conditions on the NL shelf
and in the Northwest Atlantic as a whole. It consists of annual normalized anomalies of 10 subindices with
equal contribution: Winter North Atlantic Oscillation, air temperature, sea ice, icebergs, sea surface temperature,
vertically-averaged temperature and salinity at Station~27, cold intermediate layer (CIL) core temperature at Station~27,
CIL area on 3 hydrographic sections and bottom temperature on the NL shelf.
This index runs from 1951 to 2019 and will be updated annually.
}
\author{
Jamie C. Tam
}
\keyword{datasets}
