% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{npgo}
\alias{npgo}
\title{North Pacific Gyre Oscillation (NPGO) Index (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{Tibble:
\itemize{
\item{\code{year}:} Year
\item{\code{month}:} Month
\item{\code{value}:} NPGO anomaly (index, unitless)
}
}
\item{meta}{Includes:
\itemize{
\item{\code{data_type}:} "North Pacific Gyre Oscillation Index"
\item{\code{region}:} "North Pacific"
\item{\code{location_descriptor}:} "NPGO region"
\item{\code{units}:} ""
\item{\code{source_citation}:} "UCSD/SIO; see data-raw"
\item{\code{original_value_col}:} "anomaly"
}
}
}
}
\source{
\url{http://www.o3d.org/npgo/}
}
\usage{
npgo
}
\description{
The NPGO is the second dominant mode of sea-surface height anomaly variability in the northeast Pacific, correlating with nutrient/chlorophyll fluctuations.
}
\examples{
plot(npgo)
}
\author{
Andrew Edwards
}
\keyword{datasets}
