% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\alias{grapes-or-or-grapes}
\title{Helper for NULL values}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check.}

\item{y}{Value to return if x is NULL.}
}
\value{
x if not NULL, otherwise y.
}
\description{
Returns the first argument if it is not NULL, otherwise returns the second.
}
\keyword{internal}
