% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{oni}
\alias{oni}
\title{Oceanic Niño Index (ONI)}
\format{
A tibble of class \code{pacea_index} with columns:
\describe{
\item{year}{Year of value}
\item{month}{Month (1–12) of value}
\item{val}{Three-month average SST (°C); recent values may change in subsequent updates}
\item{anom}{SST anomaly (°C); recent values may change in subsequent updates}
}
}
\source{
Generated from running \code{data-raw/coastwide-indices/coastwide-indices.R}.
}
\usage{
oni
}
\description{
The Oceanic Niño Index (ONI) is a monthly measure of the El Niño-Southern Oscillation. It is calculated as a 3-month running mean of sea surface temperature (SST) anomalies in the Niño 3.4 region (5°N–5°S, 120°W–170°W), using 30-year base periods updated every 5 years. ONI values may change for up to two months after initial posting due to data filtering.
}
\details{
For more information, see \href{http://www.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ensoyears.shtml}{NOAA ONI} and \href{https://www.ncei.noaa.gov/access/monitoring/enso/sst}{NOAA ENSO SST}.

Associated code adapted from code generously shared by Chris Rooper.
}
\examples{
\dontrun{
oni
plot(oni)
}
}
\references{
Ross, T., & Robert, M. (2022). Normal temperatures despite strong cool climate indices and an emerging freshening trend. In Boldt, J.L. et al. (Eds.), State of the physical, biological and selected fishery resources of Pacific Canadian marine ecosystems in 2021. Canadian Technical Report of Fisheries and Aquatic Sciences, 3482, vii+242 p.
}
\author{
Andrew Edwards
}
\keyword{datasets}
