% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-marea-st.R
\name{plot.marea_st}
\alias{plot.marea_st}
\title{Plot a marea spatiotemporal data layer}
\usage{
\method{plot}{marea_st}(x, ...)
}
\arguments{
\item{x}{a `marea_st` object, which is an `sf` object}

\item{...}{other arguments to be passed on, but not currently used (`?ggplot`
says the same thing); this should remove a R-CMD-check warning.}

\item{months.plot}{character or numeric vector to indicate which months to include (e.g. `c(1, 2)`, `c("April", "may")`, `c(1, "April")`)}

\item{years.plot}{vector of years to include, from 1993 to 2019}
}
\value{
plot of the spatial data to the current device (returns nothing)
}
\description{
Plot for `marea_st` classes of objects using `ggplot()`. Objects can
currently be `glorys_bottom_temperature`. Calls `pacea::plot.pacea(st)` with
Maritimes-specific defaults.
}
\details{
Gives a quick visualization of data, specifying month(s) and year(s). For
 more options and configurable plots see vignette.
}
\examples{
\dontrun{
plot(glorys_bottom_temperature)
}
}
