% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot.marea_trend}
\alias{plot.marea_trend}
\title{Plot a time series for one or all regions}
\usage{
\method{plot}{marea_trend}(
  x,
  region = NULL,
  x_lab = "Year",
  y_lab = attr(x, "axis_name"),
  mar_all_regions = c(3, 3, 2, 0),
  oma_all_regions = c(2, 1, 1, 1),
  y_tick_by = 5,
  title = "full",
  ...
)
}
\arguments{
\item{x}{A \code{marea_herring} object.}

\item{region}{The region(s) to plot. If \code{NULL}, all regions are plotted.}

\item{x_lab}{Label for the x-axis (default is "Year").}

\item{y_lab}{Label for the y-axis (default is taken from \code{attr(obj, "axis_name")}).}

\item{mar_all_regions}{Margin settings for \code{par()} when plotting all regions.}

\item{oma_all_regions}{Outer margin settings for \code{par()} when plotting all regions.}

\item{y_tick_by}{Spacing for y-axis ticks (default is 5).}

\item{title}{If \code{NULL}, no figure title is shown. If \code{"full"} (default), the full region name is used. If \code{"short"}, only the acronym is used.}

\item{...}{Additional arguments passed to \code{plot.pacea_biomass()} and \code{plot.default()}.}
}
\value{
A time series plot is displayed. Nothing is returned.
}
\description{
Create a time series plot for herring or similar data, either for a single region or all regions.
This function is adapted from \code{plot.pacea_recruitment_herring()}.
}
\examples{
\dontrun{
# plot(marea_herring_object)
}
}
\author{
Jaimie Harbin and Benoit Casault
}
