% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_red_blue}
\alias{plot_red_blue}
\title{Plot a red/blue anomaly time series (smoothed)}
\usage{
plot_red_blue(obj_lub, value, xlab, ylab, type, ...)
}
\arguments{
\item{obj_lub}{A \code{pacea_index} object with a \code{date} column (as \code{Date}).}

\item{value}{The column to plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{type}{Type of plot.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
A plot is displayed. Nothing is returned.
}
\description{
Internal function to plot a time series with red shading above zero and blue below zero, using linear interpolation for smooth transitions.
}
\examples{
\dontrun{
# see plot.pacea_index()
}
}
\author{
Andrew Edwards
}
