% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{satellite_ocean_colour}
\alias{satellite_ocean_colour}
\title{Satellite Ocean Colour (ea_spatial)}
\format{
An \code{ea_spatial} (stars) object:
\describe{
\item{t_start_value}{Day of year of the start of the spring phytoplankton bloom (day of year)}
\item{t_duration_value}{Duration of the spring phytoplankton bloom (days)}
\item{amplitude_real_value}{Maximum concentration during the spring phytoplankton bloom period (mg/m3)}
\item{magnitude_real_value}{Total chlorophyll-a produced during the spring phytoplankton bloom period (days\emph{mg/m3)}
\item{annual_mean_value}{Average chlorophyll-a over the year (mg/m3)}
}
\strong{Metadata (\code{attr(obj, "meta")})}:
\describe{
\item{data_type}{Character. "t_start: Day of year of the start of the spring phytoplankton bloom; t_duration: Duration of the spring phytoplankton bloom; amplitude_real: Maximum concentration during the spring phytoplankton bloom period; magnitude_real: Total chlorophyll-a produced during the spring phytoplankton bloom period; annual_mean: Average chlorophyll-a over the year"}
\item{units}{Character. "t_start: day of year; t_duration: days; amplitude_real: mg/m3; magnitude_real: days}mg/m3; annual_mean: mg/m3"}
\item{region}{Character. "Northwest Atlantic"}
\item{time_descriptor}{Character. Year}
\item{source_citation}{Character. "Clay, S., & Devred, E. (2025). Northwest Atlantic per-pixel spring bloom metrics derived from OC-CCI POLY4 satellite Chlorophyll-a (v1.1) \link{Data set}. https://catalogue.cioosatlantic.ca/dataset/ca-cioos_473fe40d-613e-4aba-b335-7a31a04e5d69?local=en"}
\item{temporal_coverage}{Character. "1998-2024"}
\item{variable_name}{Character. The name used for the variable}
}
}
\source{
SOPhyE group, DFO, see data-raw/satellite-ocean-colour/satellite_ocean_colour.R
}
\usage{
satellite_ocean_colour
}
\description{
Annual spring bloom metrics and average chlorophyll-a derived from the OC-CCI
POLY4 chl-a product in the Northwest Atlantic. Pulled using marea::get_erddap_data().
NRMSE_bloom > 0.5 and |standardized anomalies| > 4 filtered out. Resampled from
4.64 km to 0.1 degree resolution.
}
\author{
marea/DFO
}
\keyword{datasets}
