% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_spatial_class.R
\name{[[,ea_spatial,ANY,ANY-method}
\alias{[[,ea_spatial,ANY,ANY-method}
\title{Access Elements of an \code{ea_spatial} Object}
\usage{
\S4method{[[}{ea_spatial,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{ea_spatial} object.}

\item{i}{A character string naming an element in the \code{meta} or \code{data} slot.}

\item{...}{Additional arguments (ignored).}
}
\value{
The requested element from the \code{meta} or \code{data} slot.
}
\description{
Provides access to metadata or data layers/columns using \code{[[} syntax.
}
