% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ea_spatial.R
\name{as_ea_spatial}
\alias{as_ea_spatial}
\title{Coerce an sf object to an ea_spatial object}
\usage{
as_ea_spatial(x, value_col = NULL, ...)
}
\arguments{
\item{x}{An \code{sf} or similar object, usually with class \code{pacea_spatial}}

\item{value_col}{Name of the column to use as value. If NULL, will auto-pick if only one candidate exists.}

\item{...}{Additional metadata overrides (named elements).}
}
\value{
An \code{ea_spatial} object.
}
\description{
Convert a spatial/class \code{pacea_spatial} or \code{marea_spatial} (simple features) object or a generic sf object generic \code{ea_spatial}.
All metadata is extracted from attributes.
}
