% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_bottom_temperature}
\alias{azmp_bottom_temperature}
\title{AZMP Scotian Shelf Bottom Temperature (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character NAFO division or area name
\item{\code{value}:} Numeric temperature at sea floor (°C)
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "Annual Bottom Temperature"
\item{\code{region}:} "Scotian Shelf"
\item{\code{location_descriptor}:} "NAFO 4X-4W"
\item{\code{units}:} "°C"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, https://www.bio.gc.ca/science/monitoring-monitorage/azmp-pmza/overview-eng.php"
\item{\code{original_value_col}:} "value"
}
}
}
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdata-load.R}
}
\usage{
azmp_bottom_temperature
}
\description{
Bottom water temperature (annual mean) from the Atlantic Zone Monitoring Program (AZMP) for the Scotian Shelf, covering NAFO Divisions 4X, 4V, and 4W. Data are averaged from in situ surveys and are crucial for fisheries and climate analyses.
}
\author{
Jaimie Harbin and Benoit Casault
}
\keyword{datasets}
