% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_bottom_temperature}
\alias{azmp_bottom_temperature}
\title{AZMP Scotian Shelf Bottom Temperature}
\format{
A tibble of class \code{pacea_index} with columns:
\describe{
\item{year}{Year of value}
\item{region}{Region identifier}
\item{value}{Temperature value (°C)}
}
}
\source{
Generated from running \code{data-raw/coastwide-indices/coastwide-indices.R}.
}
\usage{
azmp_bottom_temperature
}
\description{
Average water temperatures at the sea floor, over NAFO areas 4X, 4V, and 4W, as part of the Atlantic Zone Monitoring Program (AZMP).
}
\examples{
\dontrun{
azmp_bottom_temperature
plot(azmp_bottom_temperature)
}
}
\author{
Jaimie Harbin and Benoit Casault
}
\keyword{datasets}
