% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{azmp_satellite_temperature}
\alias{azmp_satellite_temperature}
\title{AZMP Scotian Shelf SST from Satellite (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble with columns:
\itemize{
\item{\code{year}:} Numeric year
\item{\code{region}:} Character NAFO area name
\item{\code{mean_value}:} Numeric sea surface temperature (°C)
}
}
\item{meta}{A named list, including:
\itemize{
\item{\code{data_type}:} "Sea Surface Temperature"
\item{\code{region}:} "Scotian Shelf"
\item{\code{location_descriptor}:} "NAFO 4X-4W"
\item{\code{units}:} "°C"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "DFO AZMP, through azmpdata package https://github.com/casaultb/azmpdata/"
}
}
}
}
\source{
Generated from running \code{data-raw/azmpdata/azmpdata-load.R}
}
\usage{
azmp_satellite_temperature
}
\description{
Sea surface temperature (SST) from satellite data for the Scotian Shelf, covering NAFO Divisions 4X, 4V, and 4W. Data are averaged from satellite observations and are crucial for fisheries and climate analyses.
}
\author{
Emily O'Grady and Benoit Casault
}
\keyword{datasets}
