% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{coastline}
\format{
A data frame with 72 rows and 4 variables:
\describe{
  \item{\code{admin}}{character The administrative region of the coastline, e.g. "Canada", "USA", "Greenland", "St Pierre et Miquelon"}
  \item{\code{name_en}}{character The name of the province or state in English, e.g. "British Columbia", "Alaska", "Newfoundland and Labrador"}
  \item{\code{name_fr}}{character The name of the province or state in French, e.g. "Colombie-Britannique", "Alaska", "Terre-Neuve-et-Labrador"}
  \item{\code{geometry}}{list A list of geometries representing the coastline, in the form of `sf` objects. Each geometry is a polygon or a line representing #the coastline of the respective administrative region.} 
}
}
\source{
Generated from running `data-raw/coastline/coastline.R`.
}
\usage{
coastline
}
\description{
coastline of Canada, the US, Geenland, and St Pierre et Miquelon
}
\author{
Remi Daigle
}
\keyword{datasets}
