% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_data_class.R
\name{ea_data-validity}
\alias{ea_data-validity}
\title{Timeseries Data Validity}
\arguments{
\item{object}{An \code{ea_data} object.}
}
\value{
\code{TRUE} if the object is valid, otherwise a character string
describing the validation failure.
}
\description{
The validity method for the \code{ea_data} class ensures that the \code{data} slot
contains the required \code{year} and \code{value} columns. This check is performed
automatically upon object creation with \code{new()}.
}
\keyword{internal}
