% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_spatial_class.R
\name{ea_spatial-validity}
\alias{ea_spatial-validity}
\title{Validity Method for ea_spatial}
\arguments{
\item{object}{An \code{ea_spatial} object.}
}
\value{
TRUE if valid, otherwise a character string describing the issue.
}
\description{
Ensures that the \code{data} slot is a supported spatial object and contains a \code{value} layer or column.
}
\keyword{internal}
