% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{eco_indicators}
\alias{eco_indicators}
\title{ecological indicators (ea_data)}
\format{
an \code{ea_data} object with multiple columns:
\describe{
\item{year{Year of data collection}
\item{region}{assessment of management area, NAFO, Scotian Shelf}
\item{SpeciesRichness_ALL}{The number of species recorded in RV survey or commercial landings}
\item{ShannonDiveristy_ALL}{Index considering species richness and evenness of abundances}
\item{MargalefRichness_ALL}{An Index measure of richness accounting for sample size}
\item{PielouEvenness_ALL}{How evenly species are distributed within a community}
\item{HillDiversity_ALL}{effective number of species}
\item{HillDominance_ALL}{community dominance index describing how much a community's diveristy is driven bye few dominant species}
\item{Heips_ALL}{relative abundance of different species in a community}
\item{SpeciesRichness_ALL_s}{standardized species richness}
\item{ShannonDiversity_ALL_s}{sandardized Shannon diveristy index}
\item{MargalefRichness_ALL_s}{sandardized Margalef species richness}
\item{PielouEvenness_ALL_s}{standardized Pilou's evenness}
\item{HillDiversity_ALL_s}{standardized Hill diversity}
\item{HillDominance_ALL_S}{standardized Hill dominance}
\item{Heips_ALL_s}{standardized Heips}
\item{Abundance_ALL}{Total abundance}
\item{BIOMASS_ALL}{Total biomass of all species caught in RV survey (t)}
\item{BIOMASSS_CLUPEIDS}{Total biomass of clupeid species (t)}
\item{BIIOMASS_FINFISH}{Total biomass of finfish species (t) }
\item{BIOMNASS_FLATFISH}{Total biomass of flatfish species (t)}
\item{BIOMASS_FORAGE}{Total biomass of forage species (t)}
\item{BIOMASS_GADOIDS}{Total biomass of gadoids species (t)}
\item{BIOMASS_GROUNDFISH}{Total biomass of groundfish species (t)}
\item{BIOMASS_PELAGIC}{Total biomass of pelagic species (t)}
\item{BIOMASS_SKATE}{Total biomass of skate species (t)}
\item{FishinginBalance}{changes in fishing strategies and their impact on system productivity; positive FiB indicates fishery has expanded or bottom-up effects are occuring}
\item{ABUNDANCE_ALL_s}{standardized total abundance of all species caught in RV survey}
\item{BIOMASS_ALL_s}{standardized total biomass of all species caught in RV survey}
\item{BIOMASS_CLUPEIDS_s}{standardized total biomass of all clupeid species}
\item{BIOMASS_FINFISH_s}{standardized total biomass of all finsfish species}
\item{BIOMASS_FLATFISH_s}{standardized total biomass of all flatfish species}
\item{BIOMASS_FORAGE_s}{standardized total biomass of all forage species}
\item{BIOMASS_GADOIDS_s}{standardized total biomass of all gadioid species}
\item{BIOMASS_GROUNDFISH_s}{standardized total biomass of all groundfish species}
\item{BIOMASS_PELAGIC_s}{standardized total biomass of all pelagic species}
\item{BIOMASS_SKATES_s}{standardized total biomass of all akate species}
\item{FishinginBaland_s}{standardized changes in fishing strategies and their impact on system productivity}
\item{DiversityTargetSpp_ALL}{Diversity of target commercial species}
\item{MeanTL.Landings}{Mean Trophic level of all landings}
\item{MTI.Landings_3.25}{Mean Trophic Level of landings  above TL 3.25}
\item{landings_ALL}{total landings (t)}
\item{landings_CLUPEIDS.L}{total landings of clupeids species (t)}
\item{landings_FINFISH.L}{total landings of finfish species (t)}
\item{landings_FLATFISH.L}{total landings of flatfish species(t)}
\item{landings_FORAGE.L}{total landings of forage species (t)}
\item{landings_GADOIDS.L}{total landings of gadoid species (t))}
\item{landings_GROUNDFISH.L}{total landings of groundfish species (t)}
\item{landings_INVERTEBRATES.L}{total landings of invertebrate species (t)}
\item{landings_LARGE_PELAGIC.L}{total landings of pelagic species (t)}
\item{landings_SKATES.L}{total landings of skate species (t)}
\item{FP_ALL}{Fishing pressure of all species}
\item{FP_CLUPEIDS}{Fishing pressure of clupeid species}
\item{FP_FINFISH}{Fishing pressure of finfish species}
\item{FP_FLATFISH}{Fishing pressure of flatfish species}
\item{FP_FORAGE}{Fishing pressure of forage species}
\item{FP_GADOIDS}{Fishing pressure of gadoid species}
\item{FP_GROUNDFISH}{Fishing pressure of groundfish species}
\item{FP_SKATES}{Fishing pressure of skate species}
\item{DiversityTargetSpp_ALL_s}{standardized diversity of target species}
\item{MeanTL.Landings_s}{standardize mean trophic level of landings}
\item{MTI.Landings_3.25_s}{standardized mean trophic level above 3.25}
\item{landings_ALL_s}{standardized landings of all species}
\item{landings_CLUPEIDS.L_s}{standardized landings of clupeid species}
\item{landings_FINFISH.L_s}{standardized landings of finfish species (t)}
\item{landings_FLATFISH.L_s}{standardized landings of flatfish species(t)}
\item{landings_FORAGE.L_s}{standardized landings of forage species}
\item{landings_GADOIDS.L_s}{standardized landings of gadoid species}
\item{landings_GROUNDFISH.L_s}{standardized landings of groundfish species}
\item{landings_INVERTEBRATES.L_s}{standardized landings of invertebrate species}
\item{landings_LARGE_PELAGIC.L_s}{standardized landings of large pelagic species}
\item{landings_SKATES.L_s}{standardized landings of skates species}
\item{FP_ALL_s}{standardized fishing pressure of all species}
\item{FP_CLUPEIDS_s}{standardized fishing pressure of clupeid species}
\item{FP_FINFISH_s}{standardized fishing pressure of finfish species}
\item{FP_FLATFISH_s}{standardized fishing pressure of flatfish species}
\item{FP_FORAGE_s}{standardized fishing pressure of forage species}
\item{FP_GADOIDS_s}{standardized fishing pressure of gadoid species}
\item{FP_GROUNDFISH_s}{standardized fishing pressure of groundfish species}
\item{FP_SKATES_s}{standardized fishing pressure of skates species}
\item{CVBiomass}{Coefficient of Variation of the biomass}
\item{MeanLifespan}{Mean Lifespan observed for each species}
\item{BIOMASS_TL3}{biomass of trophic levle 3}
\item{BIOMASS_TL4}{biomass of trophic level 4}
\item{MMLength_BIOMASS}{Mean maxium length in the community weigthed by biomass}
\item{MMLength_ABUNDANCE}{Mean maximum length in the community weighted by abundance}
\item{IVILandings}{Intrinsic vulnerability index of the catch}
\item{CVBiomass_s}{standardized coefficient of variation of biomass}
\item{MeanLifespan_s}{standardized mean lifespan}
\item{BIOMASS_TL2_s}{standardized biomass of trophic level 2}
\item{BIOMASS_TL3_s}{standardized biomass of trophic level 3}
\item{BIOMASS_TL4_s}{standardized biomass of trophic level 4}
\item{MMLength_BIOMASS_s}{standardized mean maxium length in the community weigthed by biomass}
\item{MMLength_ABUNDANCE_s}{standardized mean maxium length in the community weigthed by abundance}
\item{IVILandings_s}{standardized Intrinsic vulnerability index of the catch}
\item{BIOMASS_LBENTHIVORE}{biomass of large benthovores (t) calculated from RV survey}
\item{BIOMASS_MBENTHIVORE}{biomass of medium size benthivores (t) calculated from RV survey}
\item{BIOMASS_PISCIVORE}{biomass of piscivores (t) calculated from RV survey}
\item{BIOMASS_PLANKTIVORE}{biomass of planktivores (t) calculated from RV survey}
\item{BIOMASS_ZOOPISCIVORE}{biomass of zoopiscivores (t) calculated from RV survey}
\item{PELAGIC_GROUNDFISH}{biomass of pelagics to biomass of groundfish}
\item{PREDATORS_ALL}{Total biomass of predators}
\item{LargeFishIndicator}{proportion of large fish > 35cm to small fish below or equal to 35cm}
\item{MeanLengthBIOMASS}{mean length of fish in the community weighted by biomass}
\item{MeanLengthABUNDANCE}{mean length of fish in the community weighted by abundance}
\item{CCondition_FINFISH}{community condition of finfish}
\item{CCondition_LBENTHIVORE}{community condition of large benthivores}
\item{CCondition_MBENTHIVORE}{community condition of medium benthivores}
\item{CCondition_PISCIVORE}{community condition of piscivores}
\item{CCondition_PLANKTIVORE}{community condition of planktivores}
\item{CCondition_ZOOPISCIVORE}{community condition of zoopiscivores}
\item{BIOMASS_LBENTHIVORE_s}{standardized biomass of large benthivores}
\item{BIOMASS_MBENTHIVORE_s}{standardized biomass of medium benthivores}
\item{BIOMASS_PISCIVORE_s}{standardized biomass of piscivores}
\item{BIOMASS_PLANKTIVORE_s}{standardized biomass of planktivores}
\item{BIOMASS_ZOOPISCIVORE_s}{standardized biomass of zoopiscivores}
\item{PELAGIC_GROUNDFISH_s}{standardized pelagic to groundfish}
\item{PREDATORS_ALL_s}{standardized predoatrs}
\item{LargeFishIndicaor_s}{standardized large fish indicator}
\item{MeanLengthBIOMASS_s}{standardized mean length biomass}
\item{MeanLengthABUNDANCE_s}{standardized mean length abundance}
\item{CCondition_FINFISH_s}{standardized community condition of finfish}
\item{CCondition_LBENTHIVORE_s}{standardized community condition of large benthivores}
\item{CCondition_MBENTHIVORE_s}{standardized community condition of medium benthivores}
\item{CCondition_PISCIVORE_s}{standardized community condition of piscivores}
\item{CCondition_PLANKTIVORE_s}{standardized community condition of planktivores}
\item{CCondition_PISCIVORE_s}{standardized community condition of piscivores}
\item{CCondition_PLANKTIVORE_s}{standardized community condition of planktivores}
\item{CCondition_ZOOPISCIVORE_s}{standardized community condition of zooplanktivores}
\item{BIOMASS_INVERTEBRATES}{biomass of invertebrates}
\item{INVERTEBRATES_GROUNDFISH}{proportion of invertebrates to groundfish}
\item{FP_INVERTEBRATES}{Fishing pressure of invertebrates}
\item{BIOMASS_TL2}{biomass of trophic level 2}
}
}
\source{
Generated from running \code{data-raw/eco-indicators/eco-indicators.R}
}
\usage{
eco_indicators
}
\description{
ecological indicators calculated from biomass estimates and commercial harvesting data
}
\author{
Jamie C. Tam
}
\keyword{datasets}
