% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{food_habits}
\alias{food_habits}
\title{Food Habits Metadata}
\format{
A data.frame with columns:
\describe{
\item{DATASOURCE}{Data source}
\item{MISSION}{Cruise/mission identifier}
\item{SETNO}{Set number}
\item{SDATE}{Date of set}
\item{STIME}{Time of set (24 hr)}
\item{STRAT}{Stratum code}
\item{BOTTOM_TEMPERATURE}{Water temperature at bottom (°C)}
\item{DEPTH}{Bottom depth (m)}
\item{GEAR}{Gear code}
\item{SLATDD}{Latitude (decimal degrees)}
\item{SLONGDD}{Longitude (decimal degrees, negative=west)}
\item{NAFO_ZONE}{NAFO zone}
\item{NAFO_SUBUNIT}{NAFO subunit}
\item{SPEC}{Species code}
\item{FSHNO}{Fish ID}
\item{FWT}{Fish weight (g)}
\item{FLEN}{Fish length (cm)}
\item{STOWGT}{Stomach weight (g)}
\item{EMPTYWGT}{Empty stomach weight (g)}
\item{FULLNESS}{Stomach fullness (coded)}
\item{FGEN}{Fish sex}
\item{PREYSPECCD}{Prey species code}
\item{PWT}{Prey weight (g)}
\item{PLEN}{Prey length (cm)}
\item{PNUM}{Number of prey}
\item{DIGESTION}{Digestive state (coded)}
}
}
\source{
Generated from running \code{data-raw/ecosystem-survey/food_habits.R}
}
\usage{
food_habits
}
\description{
Stomach content and food habits data collected during DFO ecosystem surveys in the Maritimes region. Contains trawl set, fish, and prey item information.
}
\author{
Manon Cassista Da Ros; Mike McMahon
}
\keyword{datasets}
