% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CMEMS_ncdf.R
\name{get_CMEMS_ncdf}
\alias{get_CMEMS_ncdf}
\title{Get CMEMS data}
\usage{
get_CMEMS_ncdf(
  username = NA,
  password = NA,
  dataset_id = "cmems_mod_glo_phy_my_0.083deg_P1M-m",
  variables,
  minimum_longitude = -67.7425,
  maximum_longitude = -54.90132,
  minimum_latitude = 40.04343,
  maximum_latitude = 47.83333,
  start_datetime = "1993-12-01T00:00:00",
  end_datetime = "1994-12-01T00:00:00",
  output_filename = tempfile(fileext = ".nc")
)
}
\arguments{
\item{username}{character for your CMEMS username}

\item{password}{character for your CMEMS password}

\item{dataset_id}{character string with the dataset id. Defaults to "cmems_mod_glo_phy_my_0.083deg_P1M-m" which is the global ocean physics analysis and forecast product on a 1/12° grid. Additional products can be found in the [Copernicus Data Store](https://data.marine.copernicus.eu/products)}

\item{variables}{character vector with variable names}

\item{minimum_longitude}{numeric minimum longitude of a bounding box for the data download. Defaults to -67.74250 which is for the Scotian Shelf and Bay of Fundy planning area.}

\item{maximum_longitude}{numeric maximum longitude of a bounding box for the data download. Defaults to -54.90132 which is for the Scotian Shelf and Bay of Fundy planning area.}

\item{minimum_latitude}{numeric minimum latitude of a bounding box for the data download. Defaults to 40.04343 which is for the Scotian Shelf and Bay of Fundy planning area.}

\item{maximum_latitude}{numeric maximum latitude of a bounding box for the data download. Defaults to 47.83333 which is for the Scotian Shelf and Bay of Fundy planning area.}

\item{start_datetime}{character string for the start date and time e.g. "1993-12-01T00:00:00"}

\item{end_datetime}{character string for the end date and time e.g. "1994-12-01T00:00:00"}

\item{output_filename}{character string for the output filename}
}
\description{
This function downloads Copernicus Marine Environment Monitoring Service ([CMEMS](https://marine.copernicus.eu/)) data using the Copernicus Marine API and the python package copernicusmarine.
}
