% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CMEMS_ncdf.R
\name{get_CMEMS_ncdf}
\alias{get_CMEMS_ncdf}
\title{Download Copernicus Marine (CMEMS) data}
\usage{
get_CMEMS_ncdf(
  username = NA,
  password = NA,
  dataset_id = "cmems_mod_glo_phy_my_0.083deg_P1M-m",
  variables,
  minimum_longitude = -67.7425,
  maximum_longitude = -54.90132,
  minimum_latitude = 40.04343,
  maximum_latitude = 47.83333,
  start_datetime = "1993-12-01T00:00:00",
  end_datetime = "1994-12-01T00:00:00",
  output_filename = tempfile(fileext = ".nc")
)
}
\arguments{
\item{username}{Your CMEMS username.}

\item{password}{Your CMEMS password.}

\item{dataset_id}{Dataset ID (default is "cmems_mod_glo_phy_my_0.083deg_P1M-m", the global ocean physics product).}

\item{variables}{Character vector of variable names to download.}

\item{minimum_longitude}{Minimum longitude for the data bounding box.}

\item{maximum_longitude}{Maximum longitude for the data bounding box.}

\item{minimum_latitude}{Minimum latitude for the data bounding box.}

\item{maximum_latitude}{Maximum latitude for the data bounding box.}

\item{start_datetime}{Start date and time (e.g., "1993-12-01T00:00:00").}

\item{end_datetime}{End date and time (e.g., "1994-12-01T00:00:00").}

\item{output_filename}{Name of the output file (NetCDF format).}
}
\value{
No return value. Downloads a NetCDF file to the specified location.
}
\description{
Downloads data from the Copernicus Marine Environment Monitoring Service (\href{https://marine.copernicus.eu/}{CMEMS}) using the Copernicus Marine API and the Python package \code{copernicusmarine}.
}
\examples{
\dontrun{
get_CMEMS_ncdf(username = "your_username", password = "your_password", variables = c("thetao"))
}
}
