% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{glorys_bottom_temperature}
\alias{glorys_bottom_temperature}
\title{GLORYS Bottom Temperature}
\format{
A tibble of class \code{marea_st} with columns:
\describe{
\item{year}{Year of observation}
\item{month}{Month of observation}
\item{temperature}{Bottom temperature (°C)}
\item{geometry}{Spatial geometry (sf object)}
}
}
\source{
Copernicus Marine Service (GLORYS)
}
\usage{
glorys_bottom_temperature
}
\description{
Monthly bottom temperature estimates from the GLORYS reanalysis for the Maritimes region.
}
\examples{
\dontrun{
glorys_bottom_temperature
plot(glorys_bottom_temperature)
}
}
\author{
Your Name
}
\keyword{datasets}
