% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grey_seals}
\alias{grey_seals}
\alias{grey_seals_2021}
\title{Estimated abundance of Grey Seals on the Scotian Shelf}
\format{
A tibble also of class `marea_grey_seals` with columns:
\describe{
  \item{year:}{Year of the estimate}
  \item{low:}{low end of the estimate of abundance, defined as
   the lower value of the 95\% credible interval}
  \item{median:}{median estimate of abundance, numbers of seals}
  \item{high:}{high end of the estimate of abundance, defined as
   the higher value of the 95\% credible interval}
 }

An object of class \code{marea_grey_seals} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 62 rows and 4 columns.
}
\source{
Estimates provided by Nell den Heyer, then wrangled and imported using
  `data-raw/grey-seals/grey-seals/grey-seals.R`.
}
\usage{
grey_seals

grey_seals_2021
}
\description{
TODO The Pacific Harbour Seal (\emph{Phoca vitulina richardsi}) is the most abundant pinniped species in the
Northeast Pacific and is found throughout coastal and estuarine waters of
British Columbia. A coastwide survey of harbour seals was conducted from
2015 to 2019 and used to inform the estimated trends in abundance given here.
}
\details{
TODO The Pacific Harbour Seal stock in Canada is
assessed as a single stock. The coastwide abundance of harbour
seals in British Columbia was assessed through aerial surveys conducted over multiple
years, and corrected for area covered, survey timing relative to peak
pupping, and proportion of seals hauled out during surveys. Estimated
abundance was calculated for seven regions: the Strait of Georgia (SOG),
West Coast Vancouver Island (WCVI), Queen Charlotte Strait (QCS), Discovery
Passage (DP), Central Mainland Coast (CMC), Northern Mainland Coast (NMC),
and Haida Gwaii (HG); see Figure 1 of DFO (2022) for a map showing the
regions. The coastwide estimates are also given (labelled `Coastwide` so the
plot title is capitalised), for which the means are calculated as the sum of
the regions' means and the `low` and `high` values as TODO ???


TODO The estimated abundances
were calculated using Generalised Additive Models (GAMs), and are included here
(reproducing the trends shown in Figure 3 of DFO, 2022). See that figure to see
the amount of data available in each region; these data can be added to
`pacea` if desired -- please let us know. The final year of data in each
region is saved in the object `harbour_seals_data_final_year`.

Historical estimates of abundance will change from stock assessment to stock
assessment, and so we use `grey_seals` to represent
the results from the most recent assessment, and also save
`grey_seals_2021` for the results from
the 2021. This is so that you can always refer to a specific set of
assessment results (rather than have your analyses change because you have
updated `marea` and we have replaced `grey_seals` with results from a
new assessment).

For further details see Hammill et al. (2023), TODO from which some of the above text was
adapted.

Hammill, M.O., S.P. Rossi, A. Mosnier, C.E. den Heyer,  W.D. Bowen and
G.B. Stenson. 2023.  Grey Seal Abundance and Harvest Advice in Canadian
Waters. DFO Can. Sci. Advis. Sec. Res. Doc. 2023/053. vi + 40 p.
}
\examples{
\dontrun{
grey_seals
plot(grey_seals)
}
}
\author{
Andrew Edwards and Nell den Heyer
}
\keyword{datasets}
