% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{marea_metadata}
\alias{marea_metadata}
\title{Get metadata for all datasets in marea}
\usage{
marea_metadata()
}
\value{
A data frame with dataset information
}
\description{
Get metadata for all datasets in marea
}
\examples{
marea_metadata()
}
