% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nao}
\alias{nao}
\title{NAO -- North Atlantic Oscillation Index}
\format{
A tibble also of class `pacea_index` with columns:
\describe{
  \item{year:}{year of value}
  \item{anomaly:}{Index as a measure of streghth of westerly winds over the NA}
 }
}
\source{
Generated from running `data-raw/coastwide-indices/coastwide-indices.R`.
}
\usage{
nao
}
\description{
The North Atlantic Oscillation (NAO) Index (winter), defined as the winter
(December,January, February, March) 500 mb pressure Principal Component 
Analysis which is representative of the difference between the 
Icelandic low and Azores high.
}
\details{
The NAO Index was originally defined as the 
difference in sea-level atmospheric pressures between the Azores and Iceland 
and is a measure of the strength of the westerly winds over 
the Northwest Atlantic. It represents the dominant, large-scale 
meteorological forcing over the North Atlantic Ocean. The NAO Index is 
based on a Rotated Principal Component Analysis applied to the monthly 
standardized 500 mb height anomalies, averaged over winter months of 
December through March. 

The NAO Index are based on monthly data obtained from the National 
Oceanographic and Atmospheric Administration :
https://www.ncei.noaa.gov/access/monitoring/nao/
}
\examples{
\dontrun{
nao
plot(nao)
}
}
\author{
Jamie C. Tam; Chantelle Layton
}
\keyword{datasets}
