% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{nao}
\alias{nao}
\title{North Atlantic Oscillation (NAO) Index (ea_data)}
\format{
An \code{ea_data} object:
\describe{
\item{data}{A tibble: \describe{
\item{year}{Year}
\item{value}{NAO anomaly value}
}}
\item{meta}{A named list. Keys include:
\itemize{
\item{\code{data_type}:} "NAO Index"
\item{\code{region}:} "North Atlantic"
\item{\code{location_descriptor}:} "Azores-Iceland"
\item{\code{units}:} "index"
\item{\code{species}:} \code{NA}
\item{\code{source_citation}:} "NOAA, see source"
\item{\code{original_value_col}:} e.g. "anomaly"
}
}
}
}
\source{
\url{https://www.ncei.noaa.gov/access/monitoring/nao/}
}
\usage{
nao
}
\description{
The NAO Index represents the difference in sea-level atmospheric pressures between the Azores and Iceland, calculated by PCA of 500 mb height anomalies.
}
\author{
Jamie C. Tam; Chantelle Layton
}
\keyword{datasets}
