% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-marea-trend.R
\name{plot.marea_trend}
\alias{plot.marea_trend}
\title{Plot a time series object for one or all regions}
\usage{
\method{plot}{marea_trend}(
  obj,
  region = NULL,
  x_lab = "Year",
  y_lab = attr(obj, "axis_name"),
  mar_all_regions = c(3, 3, 2, 0),
  oma_all_regions = c(2, 1, 1, 1),
  y_tick_by = 5,
  title = "full",
  ...
)
}
\arguments{
\item{obj}{a `marea_herring` object}

\item{region}{where the sample took place}

\item{x_lab}{label for x-axis (default is "Year", only shown for bottom
panel when plotting all five regions)}

\item{y_lab}{label for y-axis (default is specified by `attr(obj, "axis_name")`.}

\item{mar_all_regions}{`mar` value for `par()`}

\item{oma_all_regions}{similar to `mar_all_regions`}

\item{y_tick_by}{increment for y-axis ticks; gets
overwritten in `add_tickmarks()` if this yields more than
`y_tick_max_number` tickmarks. If using [plot.pacea_biomass()] the default
of 1 gets automatically changed to 0.25 for `plot(hake_biomass)`. If
`y_axis_reverse` is TRUE then the negative of `y_tick_by` (which should
always be positive) is passed onto `add_tickmarks()`.}

\item{title}{if `NULL` then no figure title (using `main()`), if `full` (the
default) then spell out the Stock Assessment Region, and if `short` then
just use the acronym.}

\item{...}{further options passed onto `plot.pacea_biomass()` that can
also pass onto `plot.default()`}
}
\value{
time series ggplot
}
\description{
Adapted slightly from `plot.pacea_recruitment_herring()`.
}
\examples{
\dontrun{ TODO

}
}
\author{
Jaimie Harbin and Benoit Casault
}
